+++ packages/generator/src/action-function-import/function.ts
@@ -1,13 +1,24 @@
-import { unixEOL, caps } from '@sap-cloud-sdk/util';
+import { caps, unixEOL } from '@sap-cloud-sdk/util';
 import { FunctionDeclarationStructure, StructureKind } from 'ts-morph';
 import { VdmFunctionImport, VdmServiceMetadata } from '../vdm-types';
+import { isEntityNotDeserializable } from '../edmx-to-vdm/common';
 import { getRequestBuilderArgumentsBase } from './request-builder-arguments';
+
 const parameterName = 'parameters';
 
 export function functionImportFunction(
   functionImport: VdmFunctionImport,
   service: VdmServiceMetadata
 ): FunctionDeclarationStructure {
+  const returnType = isEntityNotDeserializable(functionImport.returnType)
+    ? `Omit<FunctionImportRequestBuilder${caps(service.oDataVersion)}<${
+        functionImport.parametersTypeName
+      }, ${functionImport.returnType.returnType}>, 'execute'>`
+    : `FunctionImportRequestBuilder${caps(service.oDataVersion)}<${
+        functionImport.parametersTypeName
+      }, ${functionImport.returnType.returnType}${
+        functionImport.returnType.isCollection ? '[]' : ''
+      }>`;
   return {
     kind: StructureKind.Function,
     name: functionImport.name,
@@ -18,21 +29,26 @@ export function functionImportFunction(
         type: functionImport.parametersTypeName
       }
     ],
-    returnType: `FunctionImportRequestBuilder${caps(service.oDataVersion)}<${
-      functionImport.parametersTypeName
-    }, ${functionImport.returnType.returnType}${
-      functionImport.returnType.isCollection ? '[]' : ''
-    }>`,
-
+    returnType,
     statements: getFunctionImportStatements(functionImport, service),
     docs: [
       [
-        `${functionImport.description}${unixEOL}`,
+        getDocDescription(functionImport),
         '@param parameters - Object containing all parameters for the function import.',
-        '@returns A request builder that allows to overwrite some of the values and execute the resultng request.'
+        '@returns A request builder that allows to overwrite some of the values and execute the resulting request.'
       ].join(unixEOL)
     ]
   };
+}
+export const additionalDocForEntityNotDeserializable =
+  "The 'execute' method does not exist when using this function/action import. Please use the 'executeRaw' for getting the raw response.";
+
+function getDocDescription(functionImport: VdmFunctionImport) {
+  return `${functionImport.description} ${
+    isEntityNotDeserializable(functionImport.returnType)
+      ? additionalDocForEntityNotDeserializable
+      : ''
+  }${unixEOL}`;
 }
 
 function getFunctionImportStatements(