+++ packages/generator/src/generator.ts
@@ -1,10 +1,6 @@
 import { PathLike } from 'fs';
 import { resolve, basename, dirname } from 'path';
-import {
-  createLogger,
-  removeFileExtension,
-  splitInChunks
-} from '@sap-cloud-sdk/util';
+import { createLogger, splitInChunks } from '@sap-cloud-sdk/util';
 import { emptyDirSync } from 'fs-extra';
 import {
   Directory,
@@ -17,6 +13,12 @@ import {
 } from 'ts-morph';
 import { ModuleKind } from 'typescript';
 import { GlobSync } from 'glob';
+import {
+  getSdkMetadataFileNames,
+  getVersionForClient,
+  sdkMetaDataHeader,
+  getSdkVersion
+} from '@sap-cloud-sdk/generator-common';
 import { packageJson as aggregatorPackageJson } from './aggregator-package/package-json';
 import { readme as aggregatorReadme } from './aggregator-package/readme';
 import { batchSourceFile } from './batch/file';
@@ -49,16 +51,8 @@ import {
   functionImportSourceFile
 } from './action-function-import';
 import { enumTypeSourceFile } from './enum-type/file';
-import {
-  getSdkMetadataFileNames,
-  sdkMetaDataHeader,
-  sdkMetaDataJS
-} from './sdk-metadata/sdk-metadata';
-import {
-  getGeneratorVersion,
-  getServiceDescription,
-  getVersionForClient
-} from './sdk-metadata/pregenerated-lib';
+import { sdkMetaDataJS } from './sdk-metadata/sdk-metadata';
+import { getServiceDescription } from './sdk-metadata/pregenerated-lib';
 
 const logger = createLogger({
   package: 'generator',
@@ -132,11 +126,11 @@ export async function generateProject(
   return project;
 }
 
-function generateAggregatorPackage(
+async function generateAggregatorPackage(
   services: VdmServiceMetadata[],
   options: GeneratorOptions,
   project: Project
-): void {
+): Promise<void> {
   if (
     typeof options.aggregatorNpmPackageName !== 'undefined' &&
     typeof options.aggregatorDirectoryName !== 'undefined'
@@ -154,7 +148,7 @@ function generateAggregatorPackage(
         cloudSdkVdmHack(npmCompliantName(options.aggregatorNpmPackageName)),
         services.map(service => service.npmPackageName),
         options.versionInPackageJson,
-        getGeneratorVersion()
+        await getSdkVersion()
       ),
       options.forceOverwrite
     );
@@ -196,9 +190,9 @@ export async function generateSourcesForService(
     otherFile(
       serviceDir,
       'package.json',
-      packageJson(
+      await packageJson(
         service.npmPackageName,
-        getVersionForClient(options),
+        await getVersionForClient(options.versionInPackageJson),
         getServiceDescription(service, options),
         options.sdkAfterVersionScript
       ),
@@ -331,26 +325,30 @@ export async function generateSourcesForService(
   }
 
   if (options.generateSdkMetadata) {
-    const { clientFileName, headerFileName } = getSdkMetadataFileNames(service);
-    logger.info(`Generating sdk header metatdata ${headerFileName}...`);
+    const { clientFileName, headerFileName } = getSdkMetadataFileNames(
+      service.originalFileName
+    );
+    logger.info(`Generating sdk header metadata ${headerFileName}...`);
     const metadataDir = project.createDirectory(
       resolve(dirname(service.edmxPath.toString()), 'sdk-metadata')
     );
-    // For the file name we use the artifact.name from API which should be the unique identifier
-    const fileName = removeFileExtension(service.originalFileName);
+
     otherFile(
       metadataDir,
       headerFileName,
       JSON.stringify(
-        await sdkMetaDataHeader(fileName, getVersionForClient(options)),
+        await sdkMetaDataHeader(
+          'odata',
+          service.originalFileName,
+          options.versionInPackageJson
+        ),
         null,
         2
       ),
       options.forceOverwrite
     );
 
-    logger.info(`Generating sdk client metatdata ${clientFileName}...`);
-
+    logger.info(`Generating sdk client metadata ${clientFileName}...`);
     otherFile(
       metadataDir,
       clientFileName,