+++ packages/generator/src/sdk-metadata/sdk-metadata.spec.ts
@@ -1,27 +1,10 @@
 import nock = require('nock');
 import { createOptions } from '../../test/test-util/create-generator-options';
-import { VdmServiceMetadata } from '../vdm-types';
 import { getTestService } from './pregenerated-lib.spec';
-import {
-  getSdkMetadataFileNames,
-  sdkMetaDataHeader,
-  sdkMetaDataJS
-} from './sdk-metadata';
+import { sdkMetaDataJS } from './sdk-metadata';
 
 describe('sdk-metadata', () => {
   const service = getTestService();
-
-  it('generates the header content from strings', async () => {
-    expect(sdkMetaDataHeader('serviceName', 'clientVersion')).toMatchSnapshot();
-  });
-
-  it('generates the File names', () => {
-    expect(
-      getSdkMetadataFileNames({
-        originalFileName: 'MyService'
-      } as VdmServiceMetadata)
-    ).toMatchSnapshot();
-  });
 
   it('generates the JS metadata content for services with pregenerated lib', async () => {
     jest.spyOn(global.Date, 'now').mockImplementationOnce(() => 0);
@@ -44,7 +27,6 @@ describe('sdk-metadata', () => {
       service,
       createOptions({ versionInPackageJson: '1.0.0' })
     );
-    nock('http://registry.npmjs.org/').head(/.*/).reply(200);
 
     expect(metaData).toMatchSnapshot();
     expect(metaData.serviceStatus.status).toBe('verified');