+++ packages/openapi-generator/test/test-util.ts
@@ -1,5 +1,6 @@
 import { OpenAPIV3 } from 'openapi-types';
 import { createRefs, OpenApiDocumentRefs } from '../src/parser/refs';
+import { OpenApiDocument } from '../src/openapi-types';
 
 export const emptyDocument = {
   openapi: '3.0.0',
@@ -10,10 +11,24 @@ export const emptyDocument = {
 export function createTestRefs(
   components: OpenAPIV3.ComponentsObject = {}
 ): Promise<OpenApiDocumentRefs> {
-  return createRefs({ ...emptyDocument, components });
+  return createRefs({ ...emptyDocument, components }, { strictNaming: true });
 }
 
 export const emptyObjectSchema = {
   additionalProperties: { type: 'any' },
   properties: []
 };
+
+export const dummyOpenApiDocument: OpenApiDocument = {
+  npmPackageName: '@sap/dummy-package',
+  apis: [
+    {
+      name: 'DummyApi',
+      operations: [
+        {
+          operationId: 'DummyFunction'
+        }
+      ]
+    }
+  ]
+} as OpenApiDocument;