+++ test-packages/test-services-e2e/TripPin/microsoft-o-data-service-sample-trippin-in-memory-models-service/function-imports.d.ts
@@ -17,7 +17,7 @@ export interface GetNearestAirportParameters {
  * Get Nearest Airport.
  *
  * @param parameters - Object containing all parameters for the function import.
- * @returns A request builder that allows to overwrite some of the values and execute the resultng request.
+ * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
  */
 export declare function getNearestAirport(
   parameters: GetNearestAirportParameters