+++ test-packages/test-services/v4/multiple-schemas-service/function-imports.d.ts
@@ -9,7 +9,7 @@ export interface TestFunctionImportEntityReturnType1Parameters {}
  * Test Function Import Entity Return Type 1.
  *
  * @param parameters - Object containing all parameters for the function import.
- * @returns A request builder that allows to overwrite some of the values and execute the resultng request.
+ * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
  */
 export declare function testFunctionImportEntityReturnType1(
   parameters: TestFunctionImportEntityReturnType1Parameters
@@ -25,7 +25,7 @@ export interface TestFunctionImportEntityReturnType2Parameters {}
  * Test Function Import Entity Return Type 2.
  *
  * @param parameters - Object containing all parameters for the function import.
- * @returns A request builder that allows to overwrite some of the values and execute the resultng request.
+ * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
  */
 export declare function testFunctionImportEntityReturnType2(
   parameters: TestFunctionImportEntityReturnType2Parameters